+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## Unreleased
+
+### New Additions
+
+- `proptest::strategy::Union` and `proptest::strategy::TupleUnion` now work
+  with weighted strategies even if the sum of the weights overflows a `u32`.
+
+### Bug Fixes
+
+- Fixed values produced via `prop_recursive()` not shrinking from the recursive
+  to the non-recursive case.
+
 ## 0.5.0
 
 ### Potential Breaking Changes